Make OpenRCT2 use iconv on Linux
This makes possible to run the game with non-english locales on Linux

Fixes #2300, fixes #2268, fixes #2267

+++ src/localisation/localisation.h
@@ -46,7 +46,7 @@ void user_string_free(rct_string_id id);
 bool is_user_string_id(rct_string_id stringId);
 
 utf8 *win1252_to_utf8_alloc(const char *src);
-int win1252_to_utf8(utf8string dst, const char *src, int maxBufferLength);
+int win1252_to_utf8(utf8string dst, const char *src, size_t maxBufferLength);
 
 int rct2_to_utf8(utf8 *dst, const char *src);
 int utf8_to_rct2(char *dst, const utf8 *src);
